apps/ai-service/requirements.txt
fastapi
uvicorn
python-multipart
pypdf2
google-generativeai
python-dotenv
requests
httpx
